+++ frontend/src/components/common/Tabs/TabList.tsx
@@ -1,11 +1,7 @@
 import { useRef } from 'react';
 import { css, styled } from 'styled-components';
 
-import color from '@Styles/color';
-
-import LeftArrow from '@Assets/icons/LeftArrowIcon';
-import RightArrow from '@Assets/icons/RightArrowIcon';
-
+import TabListScrollButton from './TabListScrollButton';
 import { useTabs } from './TabsContext';
 import useTabListScroll from './hooks/useTabListScroll';
 
@@ -17,13 +13,7 @@ const TabList = () => {
 
   return (
     <Layout>
-      {hasStartScrollButton && (
-        <ScrollButton position="left">
-          <div>
-            <LeftArrow color={color.neutral[500]} onClick={() => handleMoveScroll('start')} />
-          </div>
-        </ScrollButton>
-      )}
+      {hasStartScrollButton && <TabListScrollButton position="left" moveScroll={() => handleMoveScroll('start')} />}
       <TabListContainer ref={tabList}>
         {tabs.map((tab, index) => {
           const isSelected = tab === selectedTab;
@@ -35,13 +25,7 @@ const TabList = () => {
           );
         })}
       </TabListContainer>
-      {hasEndScrollButton && (
-        <ScrollButton position="right">
-          <div>
-            <RightArrow color={color.neutral[500]} onClick={() => handleMoveScroll('end')} />
-          </div>
-        </ScrollButton>
-      )}
+      {hasEndScrollButton && <TabListScrollButton position="right" moveScroll={() => handleMoveScroll('end')} />}
     </Layout>
   );
 };
@@ -64,58 +48,6 @@ const TabListContainer = styled.ul`
   }
 
   scrollbar-width: none;
-`;
-
-type ScrollButtonProps = {
-  position: 'left' | 'right';
-};
-
-const SCROLL_BUTTON_STYLE = {
-  left: css`
-    left: -5px;
-
-    & > div {
-      justify-content: flex-start;
-      background: linear-gradient(to left, #ffffffc5, #ffffff);
-    }
-  `,
-  right: css`
-    right: -5px;
-
-    & > div {
-      justify-content: flex-end;
-      background: linear-gradient(to right, #ffffffc5, #ffffff);
-    }
-  `,
-} as const;
-
-const ScrollButton = styled.div<ScrollButtonProps>`
-  position: absolute;
-  top: 0;
-  height: 100%;
-
-  padding-bottom: 5px;
-  border-bottom: 2px solid transparent;
-
-  z-index: 5;
-
-  & > div {
-    display: flex;
-    align-items: center;
-    width: 20px;
-    height: 105%;
-  }
-
-  svg {
-    width: 16px;
-    height: 16px;
-
-    cursor: pointer;
-  }
-
-  ${({ position }) => css`
-    ${SCROLL_BUTTON_STYLE[position]}
-  `}
 `;
 
 type TabProps = {
